#1238 - ProjectUtil is too large and contains various unrelated things
