+++ webanno-webapp/src/main/java/de/tudarmstadt/ukp/clarin/webanno/webapp/dialog/AnnotationPreferenceModalPanel.java
@@ -42,7 +42,7 @@ import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationService;
 import de.tudarmstadt.ukp.clarin.webanno.api.RepositoryService;
 import de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst;
 import de.tudarmstadt.ukp.clarin.webanno.brat.annotation.BratAnnotatorModel;
-import de.tudarmstadt.ukp.clarin.webanno.brat.project.ProjectUtil;
+import de.tudarmstadt.ukp.clarin.webanno.brat.project.PreferencesUtil;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
 import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
@@ -158,7 +158,7 @@ public class AnnotationPreferenceModalPanel
                     bModel.setWindowSize(getModelObject().numberOfSentences);
                     bModel.setStaticColor(getModelObject().staticColor);
                     try {
-                        ProjectUtil.savePreference(bModel, repository);
+                        PreferencesUtil.savePreference(bModel, repository);
                     }
                     catch (FileNotFoundException e) {
                         error("Preference file not found");
