+++ webanno-webapp/src/main/java/de/tudarmstadt/ukp/clarin/webanno/webapp/page/curation/CurationPage.java
@@ -49,13 +49,14 @@ import wicket.contrib.input.events.InputBehavior;
 import wicket.contrib.input.events.key.KeyType;
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationService;
 import de.tudarmstadt.ukp.clarin.webanno.api.RepositoryService;
+import de.tudarmstadt.ukp.clarin.webanno.api.dao.SecurityUtil;
 import de.tudarmstadt.ukp.clarin.webanno.brat.annotation.BratAnnotatorModel;
 import de.tudarmstadt.ukp.clarin.webanno.brat.controller.BratAjaxCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.brat.controller.BratAnnotationException;
 import de.tudarmstadt.ukp.clarin.webanno.brat.curation.component.CurationPanel;
 import de.tudarmstadt.ukp.clarin.webanno.brat.curation.component.model.CurationBuilder;
 import de.tudarmstadt.ukp.clarin.webanno.brat.curation.component.model.CurationContainer;
-import de.tudarmstadt.ukp.clarin.webanno.brat.project.ProjectUtil;
+import de.tudarmstadt.ukp.clarin.webanno.brat.project.PreferencesUtil;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocument;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
@@ -311,7 +312,7 @@ public class CurationPage
 
                 List<SourceDocument> sourceDocumentsinIgnorState = new ArrayList<SourceDocument>();
                 for (SourceDocument sourceDocument : listOfSourceDocuements) {
-                    if (!ProjectUtil.existFinishedDocument(sourceDocument, user, repository,
+                    if (!repository.existFinishedDocument(sourceDocument, user,
                             bratAnnotatorModel.getProject())) {
                         sourceDocumentsinIgnorState.add(sourceDocument);
                     }
@@ -367,7 +368,7 @@ public class CurationPage
 
                 List<SourceDocument> sourceDocumentsinIgnorState = new ArrayList<SourceDocument>();
                 for (SourceDocument sourceDocument : listOfSourceDocuements) {
-                    if (!ProjectUtil.existFinishedDocument(sourceDocument, user, repository,
+                    if (!repository.existFinishedDocument(sourceDocument, user,
                             bratAnnotatorModel.getProject())) {
                         sourceDocumentsinIgnorState.add(sourceDocument);
                     }
@@ -1023,7 +1024,7 @@ public class CurationPage
             bratAnnotatorModel.initForDocument(mergeJCas);
 
             // Load user preferences
-            ProjectUtil.setAnnotationPreference(username, repository, annotationService,
+            PreferencesUtil.setAnnotationPreference(username, repository, annotationService,
                     bratAnnotatorModel, Mode.CURATION);
             
             // if project is changed, reset some project specific settings
